Move downloads to a separate app.
